+++ CHANGELOG.md
@@ -25,6 +25,7 @@ Dropping a requirement of a major version of a dependency is a new contract.
 
 ### Added
 - `SshInstance` now contains AWS SDK `Instance` object, which includes i.a. information about private IP address. Unblocks [JPERF-730].
+- `ProvisionedStack` now allows to find its security groups based on provided logical id. Unblocks [JPERF-730].
 
 ## [1.7.1] - 2021-09-28
 [1.7.1]: https://bitbucket.org/atlassian/aws-resources/branches/compare/release-1.7.1%0Drelease-1.7.0